Ft_Transcendence/backend/requirements.txt
Django
Pillow
djangorestframework
djangorestframework-simplejwt
django-cors-headers
psycopg2-binary
environs
channels
channels_redis
drf-spectacular
django-allauth
requests_oauthlib
argon2-cffi
requests
Daphne
asgiref
qrcode
pyotp
olefile